feat(tests): add comprehensive coverage for scalar DSL functions and utilities
- Introduced a new test file `cov_scalars.rs` to cover various scalar DSL functions.
- Added tests for numeric, string, datetime, and JSON functions to ensure proper execution.
- Enhanced spatial DSL tests to include distance and area calculations.
- Added roundtrip tests for remaining integer newtypes and bit/bignum newtypes.
- Improved macro attribute parsing with detailed tests for scalar, table, cast, and aggregate attributes.
- Added tests for deriving DuckEnum and DuckStruct, ensuring proper handling of renaming and error cases.
- Implemented tests for aggregate functions, validating structure and behavior.

+++ crates/better-duck-diesel/src/expressions/aggregates/approximate.rs
@@ -3,7 +3,7 @@
 //! Docs: <https://duckdb.org/docs/current/sql/functions/aggregates>
 
 use diesel::expression::functions::define_sql_function;
-use diesel::sql_types::{Double, Text};
+use diesel::sql_types::{Double, Float, Text};
 
 define_sql_function! {
     /// `approx_count_distinct(x)` — HyperLogLog cardinality estimate.
@@ -12,8 +12,10 @@ define_sql_function! {
 }
 define_sql_function! {
     /// `approx_quantile(x, pos)` — T-Digest quantile estimate at `pos` in `[0,1]`.
+    ///
+    /// DuckDB types the position argument as `FLOAT`, not `DOUBLE`.
     #[aggregate]
-    fn approx_quantile(x: Double, pos: Double) -> diesel::sql_types::Nullable<Double>;
+    fn approx_quantile(x: Double, pos: Float) -> diesel::sql_types::Nullable<Double>;
 }
 define_sql_function! {
     /// `reservoir_quantile(x, q)` — reservoir-sampled quantile estimate.
